Moved schema getter to the hook
